+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 - Toggle doors between secret and normal with ctrl+click
 - Makes the size of door controls independent of the scene's grid size
 
+### Bugfixes
+- In cloned scenes Locked Door Alerts will now only highlight the door in the correct scene
+
 ## v1.0.1
 - When adding a door to a synchronization group adjust it's state to bring it in sync with the other doors
 - Use the players character as speaker for the Locked Door Alert